python-open-library-mcp/pyproject.toml
[project]
name = "open-library-mcp"
version = "0.1.0"
description = "A simple FastMCP server for querying Open Library API"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
